requirements.txt
# Core LLM and agent orchestration
langchain
langchain-openai
langchain-google-genai
langgraph
langsmith

# Google and OpenAI SDKs
openai
google-generativeai
google-api-core
google-auth

# Environment/config
python-dotenv

# Data validation and settings
pydantic
pydantic-settings

# Database (if needed)
SQLAlchemy

# Utilities
requests
rich

# Optional: For async, tracing, and other common needs
tenacity
tiktoken